chore: fix post patch application
